Recreated GUI from scratch using OOP features

+++ SolutionChecker/src/problems.h
@@ -81,7 +81,7 @@ public:
 
 	std::vector<Problem> GetProblemList() { return m_problem_list; }
 	Problem GetProblem(unsigned int problem_id_) { 
-		problem_id_ = min(problem_id_, m_problem_list.size() - 1);
+		problem_id_ = min(problem_id_, (unsigned int)(m_problem_list.size() - 1));
 		return m_problem_list[problem_id_]; 
 	}
